Adding Update Query

+++ Views/DisplayProductsAdmin.php
@@ -9,29 +9,19 @@ include "../connection.php";
 include "../validation.php";
 $error_add;
 $error_update;
+
+//*Delete the Product
 if (isset($_GET['delete_id']) && !empty($_GET['delete_id'])) {
     var_dump($_GET['delete_id']);
     DeleteProductQuery($_GET['delete_id']);
 }
 
-if ($_SERVER['REQUEST_METHOD'] === 'POST'  && !empty($_POST)) {
-    
-    if ($_GET['action'] === 'add') {
-        // AddProductQuery($_POST);
-      // Handle add form data here
-      $error_add= $error['name'];
-
-    } else if ($_GET['action'] === 'update') {
-        $category_id = $_GET['category_id'];
-        UpdateCategory($category_id, $_POST);
-        // $error_update= $error['name'];
-      // Handle update form data here
-    }
-  }
 ?>
 
 <div class="container w-50">
-    <h1 class="text-primary mx-auto w-50">Products</h1>
+    <h1 class="text-primary mx-auto w-50 my-4">Products</h1>
+    <a class="btn btn-primary" href="Add Products.php">Add Product</a>
+
     <table class="table table-striped table-dark border-light  text-center table-bordered">
         <thead>
             <tr>
@@ -40,33 +30,40 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST'  && !empty($_POST)) {
                 <th>price</th>
                 <th>Category Name</th>
                 <th>quantity</th>
+                <th>Status</th>
                 <th>image</th>
                 <th>Edit</th>
                 <th>Delete</th>
             </tr>
         </thead>
-       <tbody>
+        <tbody>
             <?php
             $Products = DisplayProductsQuery();
             // var_dump(   $Products );
             foreach ($Products as $row) {
             ?>
-            <tr>
-               
-                <td><?= $row['id'] ?></td>
-                <td><?= $row['name'] ?></td>
-                <td><?= $row['price'] ?></td>
-                <td><?= SelectCategoryByIdQuery( $row['category_id'] )   ?></td>
-                <td><?= $row['quantity'] ?></td>
-                <td> <img height="100"  width="100" src="../uploads/<?= $row['image'] ?>" alt=""> </td>
-              
-                <td><a href="" class="btn btn-warning edit-category" data-toggle="modal" data-target="#modelId"
-                        data-category-id="<?= $row['id'] ?>">Edit</a></td>
-                <td><a href="?delete_id=<?= $row['id'] ?>" class="btn btn-danger"
-                        onclick="return confirm('Are you sure you want to delete this category?')">Delete</a></td>
-                        <?php     }   ?>
-            </tr>
-        </tbody>  
+                <tr>
+
+                    <td><?= $row['id'] ?></td>
+                    <td><?= $row['name'] ?></td>
+                    <td><?= $row['price'] ?></td>
+                    <td><?= SelectCategoryByIdQuery($row['category_id'])   ?></td>
+                    <td><?= $row['quantity'] ?></td>
+                    <td><?php
+                        if ($row['quantity'] <= 0) {
+                            echo "Not Available";
+                        } else {
+                            echo "Available";
+                        }
+
+                        ?></td>
+                    <td> <img height="100" width="100" src="../uploads/<?= $row['image'] ?>" alt=""> </td>
+
+                    <td><a href="UpdateProducts.php?product_id=<?= $row['id'] ?>" class="btn btn-warning edit-category">Edit</a></td>
+                    <td><a href="?delete_id=<?= $row['id'] ?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this category?')">Delete</a></td>
+                <?php     }   ?>
+                </tr>
+        </tbody>
     </table>
 </div>
